Move to pnpm workspace
Update CI workflow
