+++ .npmrc
@@ -1,0 +1,1 @@
+strict-peer-dependencies=false
