--- demo/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-    <title>Nuka Carousel Demo</title>
-    <meta name="description" content="" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
-    />
-  </head>
-  <body>
-    <!--[if lt IE 8]>
-      <p class="browsehappy">
-        You are using an <strong>outdated</strong> browser. Please
-        <a href="http://browsehappy.com/">upgrade your browser</a> to improve
-        your experience.
-      </p>
-    <![endif]-->
-    <div id="content">
-      <h1>
-        If you can see this, something is broken (or JS is not enabled)!!.
-      </h1>
-    </div>
-    <script type="text/javascript" src="/assets/main.js"></script>
-  </body>
-</html>
