+++ examples/nextjs/.eslintrc
