+++ examples/nextjs/styles/Home.module.css
@@ -4,7 +4,6 @@
 }
 
 .main {
-  min-height: 100vh;
   padding: 4rem 0;
   max-width: 1000px;
   margin: 0 auto;
