+++ package.json
@@ -1,105 +1,40 @@
 {
-  "name": "nuka-carousel",
-  "version": "5.1.2",
-  "description": "Pure React Carousel",
-  "main": "lib/index.js",
-  "module": "es/index.js",
-  "jsnext:main": "es/index.js",
-  "types": "types/index.d.ts",
-  "engines": {
-    "node": ">=12.0.0"
+  "name": "nuka-carousel-monorepo",
+  "scripts": {
+    "build": "pnpm run --filter nuka-carousel build",
+    "build-storybook": "pnpm run --filter nuka-carousel storybook:build",
+    "lint": "pnpm run --parallel lint",
+    "lint:fix": "pnpm run --parallel lint --fix",
+    "start:nextjs": "pnpm run --filter nuka-carousel-nextjs-example dev",
+    "start:storybook": "pnpm run --filter nuka-carousel storybook",
+    "test": "pnpm --filter nuka-carousel test:e2e",
+    "test:ci": "pnpm --filter nuka-carousel test:e2e:ci",
+    "test:ci-with-server": "start-server-and-test start:nextjs http://localhost:3000 test:ci",
+    "check": "pnpm run lint",
+    "check:ci": "pnpm run check",
+    "clean": "pnpm run -r clean",
+    "changeset": "changeset",
+    "version": "pnpm changeset version && pnpm install --fix-lockfile"
   },
-  "dependencies": {},
   "devDependencies": {
-    "@babel/cli": "^7.10.4",
-    "@babel/core": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.10.4",
-    "@babel/plugin-transform-object-assign": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.4",
-    "@babel/plugin-transform-typescript": "^7.16.8",
-    "@babel/polyfill": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
-    "@storybook/builder-webpack5": "^6.4.22",
-    "@storybook/manager-webpack5": "^6.4.22",
-    "@storybook/react": "^6.4.22",
-    "@types/exenv": "^1.2.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.5.0",
-    "@typescript-eslint/parser": "^5.5.0",
-    "babel-loader": "8.1.0",
-    "builder": "^5.0.0",
-    "chromatic": "^6.2.0",
+    "@babel/core": "^7.17.2",
+    "@changesets/cli": "^2.23.1",
+    "@types/jest": "^27.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "babel-eslint": "^10.1.0",
     "cypress": "^9.3.1",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.15.6",
     "eslint": "^8.14.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.5.0",
-    "exenv": "^1.2.0",
-    "prettier": "^2.6.2",
-    "prop-types": "^15.6.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-move": "^6.5.0",
-    "typescript": "^4.5.2",
-    "url-loader": "^4.1.1",
-    "webpack": "5.61.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.6.0"
-  },
-  "peerDependencies": {
-    "react": "^15.3.0 || ^16.0.0 || ^17.0.1 || ^18.0.0",
-    "react-dom": "^15.3.0 || ^16.0.0 || ^17.0.1 || ^18.0.0"
-  },
-  "resolutions": {
-    "jsdom": "^14.0.0"
-  },
-  "scripts": {
-    "build": "builder concurrent --buffer build-lib build-es build-ts",
-    "build-babel": "babel src --extensions .ts,.tsx",
-    "build-es": "builder run  --env \"{\\\"BABEL_ENV\\\":\\\"esm\\\"}\" build-babel -- -d es",
-    "build-lib": "builder run build-babel -- -d lib",
-    "build-ts": "tsc --project tsconfig.build.json",
-    "build-watch": "builder concurrent build-lib build-es -- --watch",
-    "check": "npm run lint && npm run check-typescript",
-    "check-typescript": "tsc --noEmit",
-    "lint": "eslint --ext .js,.ts,.tsx .",
-    "lint-fix": "eslint --ext .js,.ts,.tsx . --fix",
-    "prettier": "prettier \"**/*.{js,json,ts,tsx,css,md}\"",
-    "prettier-fix": "prettier \"**/*.{js,json,ts,tsx,css,md}\" --write",
-    "preversion": "npm run check",
-    "start": "webpack-dev-server --mode development",
-    "start:next": "cd ./examples/nextjs && yarn dev",
-    "start:next:ci": "cd ./examples/nextjs && yarn install && yarn dev",
-    "test:e2e": "cypress open",
-    "test:e2e:ci": "cypress run",
-    "version": "npm run build",
-    "storybook": "start-storybook -p 6006",
-    "storybook-build": "build-storybook",
-    "chromatic": "chromatic --exit-zero-on-changes"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/FormidableLabs/nuka-carousel.git"
-  },
-  "keywords": [
-    "react",
-    "carousel",
-    "nuka"
-  ],
-  "author": "@FormidableLabs",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/FormidableLabs/nuka-carousel/issues"
-  },
-  "homepage": "https://github.com/FormidableLabs/nuka-carousel",
-  "sideEffects": false
+    "start-server-and-test": "^1.14.0",
+    "prettier": "^2.4.1",
+    "pretty": "^2.0.0",
+    "typescript": "^4.5.2"
+  }
 }
