+++ packages/nuka/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../.babelrc"
+}
