+++ packages/nuka/.eslintignore
@@ -1,5 +1,4 @@
 node_modules
 lib
 es
-examples
 types
