+++ packages/nuka/cypress/integration/infinity-carousel.spec.js
@@ -257,203 +257,6 @@ context('Infinity Carousel', () => {
         .should('have.attr', 'data-slide', 'Slide 3');
     });
 
-    it('should render infinity carousel with 4 slides and 3 visible slides and go through all of the slides with autoplay and without clicking the Next button', () => {
-      const params = {
-        slidesToShow: 3,
-        autoplay: true,
-        autoplayInterval: 1000,
-        wrapAround: true
-      };
-
-      cy.visit(
-        `http://localhost:3000/?slides=4&params=${JSON.stringify(params)}`
-      );
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 1');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 2');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 3');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 4');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 1');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 2');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 3');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 4');
-
-      cy.wait(1100);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 1');
-    });
-
-    it('should render infinity carousel with 4 slides and 3 visible slides and go through all of the slides starting from the last slide with autoplay and without clicking the Prev button', () => {
-      const params = {
-        slidesToShow: 3,
-        autoplay: true,
-        autoplayInterval: 3000,
-        autoplayReverse: true,
-        wrapAround: true
-      };
-
-      cy.visit(
-        `http://localhost:3000/?slides=4&params=${JSON.stringify(params)}`
-      );
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 4');
-
-      cy.wait(3050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 3');
-      cy.wait(3050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 2');
-
-      cy.wait(3500);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 1');
-
-      cy.wait(3050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 4');
-
-      cy.wait(3050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .last()
-        .should('have.attr', 'data-slide', 'Slide 3');
-    });
-
-    it('should render infinity carousel with 5 slides and 3 visible slides and go through all of the slides with autoplay and without clicking the Next button and when hover the slider should pause', () => {
-      const params = {
-        slidesToShow: 3,
-        autoplay: true,
-        autoplayInterval: 1000,
-        pauseOnHover: true,
-        wrapAround: true
-      };
-
-      cy.visit(
-        `http://localhost:3000/?slides=5&params=${JSON.stringify(params)}`
-      );
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 1');
-
-      cy.wait(1050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 2');
-
-      cy.get('.slider-container').trigger('mouseover');
-      cy.wait(1500);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 2');
-
-      cy.get('.slider-container').trigger('mouseout');
-
-      cy.wait(1050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 3');
-
-      cy.wait(1050);
-
-      cy.get('.slide.slide-visible')
-        .should('have.length', 3)
-        .find('img')
-        .first()
-        .should('have.attr', 'data-slide', 'Slide 4');
-    });
-
     it('should render infinity carousel with 4 slides and all of them should be visible and go through all of the slides', () => {
       const params = {
         slidesToShow: 4,
