+++ packages/nuka/cypress/integration/swipe-carousel.spec.js
