+++ packages/nuka/src/default-carousel-props.tsx
