+++ packages/nuka/src/utils.ts
@@ -33,10 +33,6 @@ export const addEvent = (
   }
   if (elem.addEventListener) {
     elem.addEventListener(type, eventHandler, false);
-  } else if (elem.attachEvent) {
-    elem.attachEvent(`on${type}`, eventHandler);
-  } else {
-    elem[`on${type}`] = eventHandler;
   }
 };
 
@@ -50,10 +46,6 @@ export const removeEvent = (
   }
   if (elem.removeEventListener) {
     elem.removeEventListener(type, eventHandler, false);
-  } else if (elem.detachEvent) {
-    elem.detachEvent(`on${type}`, eventHandler);
-  } else {
-    elem[`on${type}`] = null;
   }
 };
