--- tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "noEmit": false,
-    "outDir": "types"
-  },
-  "include": [
-    "globals.d.ts",
-    "src/**/*.ts",
-    "src/**/*.tsx",
-  ],
-}
